packages/python/pyproject.toml
[build-system]
requires = ["maturin>=1.13,<2.0"]
build-backend = "maturin"

[project]
name = "mdm-core"
version = "0.1.0"
description = "High-performance HWP/HWPX/PDF/DOCX to Markdown converter (Rust-powered)"
license = {text = "MIT"}
requires-python = ">=3.8"
keywords = ["markdown", "hwp", "pdf", "docx", "converter", "korean"]
classifiers = [
    "Programming Language :: Rust",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: 3",
    "Topic :: Text Processing :: Markup :: Markdown",
]

[project.optional-dependencies]
langchain = ["langchain-core>=0.1"]
llamaindex = ["llama-index-core>=0.10"]
all = ["langchain-core>=0.1", "llama-index-core>=0.10"]

[tool.maturin]
features = ["pyo3/extension-module"]
python-source = "python"
module-name = "mdm_core._mdm_native"
